♻️  Extract _exec function into new src/utils.js file
